refactor: storage 파일 utils 폴더로 이동함
